fix: remove the global flag in `io.py`
Currently, we use the return value of `execute_command` to handle the
case user editing their messages using `!editor` command. In previous
versions of the tool, `execute_command` would return the command string
directly (and the value is not used), but now we set default return value
to empty string except the `!editor` command, which will return the text
string that user edited.