feat: add black and mdformat to CI
The CI will automatically check the formatting of python files with
black and the format of markdown files with mdformat. The CI will fail
if there are any format issues.

In order to format files on you local branch, run

```
black .
mdformat .
```

Note that this will happen **in-place**. You can use the `--check` flag
to see if any files need to be formatted.