fix: remove option `wrap` in `.mdformat.toml`
Current `mdformat` causes the READMEs format to be changed in a way not
as expected. It causes the layout of long paragraphs to be broken into
multiple lines and the alignment of the links at the first line of READMEs
to be broken. This is the result of the option `wrap = 90` in `.mdformat.toml`.

This commit remove this option and revert the changes