feat: add `code_theme` to markdown and config.yaml
The user can now change the theme inside of `config.yaml`. Note that
this commit also includes a change in the `load_config`. We are testing
to see if the config is valid or not (does it contain all the needed
data).

The current implementation uses recursive functions. I think the
implementation is okay as is, though I can understand any concern
regarding recursive functions.